pyproject.toml
[tool.poetry]
name = "inverse-projection"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
notebook = "^7.2.2"
ipykernel = "^6.29.5"
numpy = "^2.0"
scikit-learn = "^1.5.2"
pyqt6 = "^6.7.1"
torch = "^2.5.0"
numba = "^0.60.0"
umap-learn = "^0.5.6"
seaborn = "^0.13.2"
shiny = "^1.2.1"
plotly = "^5.24.1"
faiss-cpu = "^1.9.0.post1"
google-api-python-client = "^2.159.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
snc = "^0.2.1"
optuna = "^4.2.0"
yellowbrick = "^1.5"
matplotlib = "^3.10.0"
scikit-image = "^0.25.2"
shapely = "^2.1.0"
asttokens = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
